requirements.txt
pynacl
discord.py
python-dotenv
